backend-api/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
pydantic==1.10.8
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
azure-storage-blob==12.16.0
PyPDF2==3.0.1
openai==1.12.0
uvicorn==0.22.0
json_repair==0.*